SealScriptInfer_V1.0/requirements.txt
torch
torchvision
timm
Pillow
opencv-python
numpy
fonttools
rarfile
opencc-python-reimplemented
albumentations
tqdm
